feat: Enhance Habits Management with Search, Dialog, and Persistence
- Replaced `LiveData` with `StateFlow` for `showDialog` and `searchQuery` in `HabitsViewModel` for better state management.
- Introduced `filterList` `StateFlow` to manage filtered habits list.
- Implemented real-time filtering in `HabitsViewModel` using `getHabitsByTitle`.
- Refactored `onSearchQueryChanged` to use `update` for reactive updates.
- Updated `onDialogClose` and `onShowDialogSelected` to use `update` for better state management.
- Modified `onTaskCreated` to accept title and description.
- Updated `DialogPopUp` to be displayed always and handle `onDismiss`.
- Improved `DialogPopUp` layout and button handling.
- Updated `HabitsScreen` to use `collectAsStateWithLifecycle`.
- Implemented `HabitsContent` to filter habits based on search query.
- Removed `LaunchedEffect` and unused variable on `HabitsContent`.
- Updated `HabitsRepository` to fetch habits based on the title.
- Modified `HabitsRepository` to use a new function to add a habit.
- Refactored `addHabit` in `HabitsRepository` to not verify if exist.
- Updated `Habits.toEntity` to use title and description.
- Modified `HabitsEntity` to use `Long` instead of `Int` for `id` field.
- Updated `Habits` to use `Long` instead of `Int` for `id` field.
- Updated `HabitsDao` to add `getHabitsByTitle` function.
- Updated `HabitsDao.getHabitById` function to get the habit by `Int`.
- Fixed minor issue on `CustomTopAppBar`.
- Updated `build.gradle.kts` to add `buildConfig` property.

+++ app/src/main/java/com/mundocode/pomodoro/model/room/HabitsEntity.kt
@@ -6,7 +6,7 @@ import com.mundocode.pomodoro.core.room.PomodoroDatabase.Companion.HABITS_TABLE_
 
 @Entity(tableName = HABITS_TABLE_NAME)
 data class HabitsEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0, // ✅ Auto-generar IDs únicos en Room
+    @PrimaryKey(autoGenerate = true) val id: Long = 0, // ✅ Auto-generar IDs únicos en Room
     val title: String,
     val description: String,
 )
